Added export alar2 command
