+++ src/JUS.CLI/JUS/ContainerCommands.cs
@@ -1,0 +1,58 @@
+// Copyright (c) 2022 Pablo Rivero
+
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+using System;
+using System.IO;
+using JUSToolkit.Containers.Converters;
+using Yarhl.FileSystem;
+
+namespace JUSToolkit.CLI.JUS
+{
+    /// <summary>
+    /// Commands related to container files.
+    /// </summary>
+    public static class ContainerCommands
+    {
+        /// <summary>
+        /// Export all the files from the Alar2 container.
+        /// </summary>
+        /// <param name="container">The path to the alar2 file.</param>
+        /// <param name="output">The output directory.</param>
+        public static void ExportAlar2(string container, string output)
+        {
+            Node files = NodeFactory.FromFile(container)
+                .TransformWith<BinaryAlar2Container>();
+
+            if (files is null) {
+                throw new FormatException("Invalid container file");
+            }
+
+            foreach (var node in Navigator.IterateNodes(files)) {
+                Console.WriteLine(node.Path);
+                Console.WriteLine(node.Name);
+                if (!node.IsContainer) {
+                    string outputFile = Path.Combine(output, node.Path);
+                    Console.WriteLine(outputFile);
+                    node.Stream.WriteTo(outputFile);
+                }
+            }
+            
+        }
+    }
+}
