+++ src/JUS.CLI/JUS/GraphicCommands.cs
@@ -1,0 +1,97 @@
+// Copyright (c) 2022 Pablo Rivero
+
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+using System;
+using System.IO;
+using JUSToolkit.Containers.Converters;
+using JUSToolkit.Graphics;
+using JUSToolkit.Graphics.Converters;
+using Yarhl.FileSystem;
+using Texim.Formats;
+using Texim.Images;
+using Texim.Sprites;
+
+namespace JUSToolkit.CLI.JUS
+{
+    /// <summary>
+    /// Commands related to graphics files.
+    /// </summary>
+    public static class GraphicCommands
+    {
+        /// <summary>
+        /// Export a DTX into PNG komas.
+        /// </summary>
+        /// <param name="container">The koma.aar container.</param>
+        /// <param name="koma">The koma.bin file.</param>
+        /// <param name="kshape">The kshape.bin file.</param>
+        /// <param name="output">The output folder.</param>
+        public static void Export(string container, string koma, string kshape, string output)
+        {
+            Node images = NodeFactory.FromFile(container)
+                .TransformWith<BinaryAlar2Container>()
+                .Children["koma"];
+            if (images is null) {
+                throw new FormatException("Invalid container file");
+            }
+
+            var shapes = NodeFactory.FromFile(kshape)
+                .TransformWith<BinaryKShape2SpriteCollection>()
+                .GetFormatAs<KShapeSprites>();
+
+            Koma komaFormat = NodeFactory.FromFile(koma)
+                .TransformWith<Binary2Koma>()
+                .GetFormatAs<Koma>();
+            foreach (KomaElement komaElement in komaFormat) {
+                string filename = $"{komaElement.KomaName}.dtx";
+
+                Node dtx = images.Children[filename];
+                if (dtx is null) {
+                    Console.WriteLine("- Missing: " + filename);
+                    continue;
+                }
+
+                dtx.TransformWith<BinaryDstx2SpriteImage>();
+                var image = dtx.Children["image"].GetFormatAs<IndexedPaletteImage>();
+
+                // We ignore the sprite info from the DSTX and we take the one
+                // from the kshape
+                var sprite = shapes.GetSprite(komaElement.KShapeGroupId, komaElement.KShapeElementId);
+
+                string outputFilePath = Path.Combine(
+                    output,
+                    $"{komaElement.KShapeGroupId}",
+                    komaElement.KomaName + ".png");
+
+                var spriteParams = new Sprite2IndexedImageParams {
+                    RelativeCoordinates = SpriteRelativeCoordinatesKind.TopLeft,
+                    FullImage = image,
+                };
+                var indexedImageParams = new IndexedImageBitmapParams {
+                   Palettes = image,
+                };
+                new Node("sprite", sprite)
+                    .TransformWith<Sprite2IndexedImage, Sprite2IndexedImageParams>(spriteParams)
+                    .TransformWith<IndexedImage2Bitmap, IndexedImageBitmapParams>(indexedImageParams)
+                    .Stream.WriteTo(outputFilePath);
+            }
+
+            Console.WriteLine("Done!");
+        }
+    }
+}
